JSON を extlib に

+++ tools/build_extlib.sh
@@ -3,26 +3,27 @@
 script=`readlink -f $0`
 root=`dirname $script`/..
 
-cd $root
-echo Setup local::lib
-mkdir tmp
-cd tmp
-wget http://search.cpan.org/CPAN/authors/id/A/AP/APEIRON/local-lib-1.004009.tar.gz
-tar xzvf local-lib*.tar.gz
-cd local-lib*
-perl Makefile.PL --bootstrap=$root/extlib --no-manpages
-make
-make install
-cd ..
-cd ..
-rm -rf tmp
+if [ "$1" = "--init" ]; then
+	cd $root
+	echo Setup local::lib
+	mkdir tmp
+	cd tmp
+	wget http://search.cpan.org/CPAN/authors/id/A/AP/APEIRON/local-lib-1.004009.tar.gz
+	tar xzvf local-lib*.tar.gz
+	cd local-lib*
+	perl Makefile.PL --bootstrap=$root/extlib --no-manpages
+	make
+	make install
+	cd ..
+	cd ..
+	rm -rf tmp
+fi
 
 cd $root
 echo Install deps to $root
 eval `perl -I$root/extlib/lib/perl5 -Mlocal::lib=$root/extlib`
 
 perl tools/cpanm --notest parent
-perl tools/cpanm --notest Try::Tiny
 perl tools/cpanm --notest Try::Tiny
 perl tools/cpanm --notest Devel::StackTrace
 perl tools/cpanm --notest Devel::StackTrace::AsHTML
@@ -36,4 +37,5 @@ perl tools/cpanm --notest Text::MicroMason
 perl tools/cpanm --notest Text::MicroMason::SafeServerPages
 perl tools/cpanm --notest Class::Factory::Util
 perl tools/cpanm --notest Class::MixinFactory
+perl tools/cpanm --notest JSON
